Refactor to create a common Fragment superclass for EventFragment and MainFragment.  Move common functions to the new superclass.  See video explanation, here: http://somup.com/cYeUF51Fsg
