+++ app/src/main/java/edu/uc/jonesbr/myplantdiary/ui/main/DiaryFragment.kt
@@ -1,0 +1,84 @@
+package edu.uc.jonesbr.myplantdiary.ui.main
+
+import android.Manifest
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.net.Uri
+import android.os.Environment
+import android.provider.MediaStore
+import android.widget.Toast
+import androidx.core.content.ContextCompat
+import androidx.core.content.FileProvider
+import androidx.fragment.app.Fragment
+import java.io.File
+import java.text.SimpleDateFormat
+import java.util.*
+
+open class DiaryFragment : Fragment() {
+    protected val SAVE_IMAGE_REQUEST_CODE: Int = 1999
+    protected val CAMERA_REQUEST_CODE: Int = 1998
+    protected val CAMERA_PERMISSION_REQUEST_CODE = 1997
+    private lateinit var currentPhotoPath: String
+    protected var photoURI : Uri? = null
+    /**
+     * See if we have permission or not.
+     */
+    protected fun prepTakePhoto() {
+        if (ContextCompat.checkSelfPermission(context!!, Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED) {
+            takePhoto()
+        } else {
+            val permissionRequest = arrayOf(Manifest.permission.CAMERA);
+            requestPermissions(permissionRequest, CAMERA_PERMISSION_REQUEST_CODE)
+
+        }
+    }
+
+    protected fun takePhoto() {
+        Intent(MediaStore.ACTION_IMAGE_CAPTURE).also{
+                takePictureIntent -> takePictureIntent.resolveActivity(context!!.packageManager)
+            if (takePictureIntent == null) {
+                Toast.makeText(context, "Unable to save photo", Toast.LENGTH_LONG).show()
+            } else {
+                // if we are here, we have a valid intent.
+                val photoFile: File = createImageFile()
+                photoFile?.also {
+                    photoURI = FileProvider.getUriForFile(activity!!.applicationContext, "com.myplantdiary.android.fileprovider", it)
+                    takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, photoURI)
+                    startActivityForResult(takePictureIntent, SAVE_IMAGE_REQUEST_CODE)
+                }
+            }
+        }
+    }
+
+    private fun createImageFile() : File {
+        // genererate a unique filename with date.
+        val timestamp: String = SimpleDateFormat("yyyyMMdd_HHmmss").format(Date())
+        // get access to the directory where we can write pictures.
+        val storageDir: File? = context!!.getExternalFilesDir(Environment.DIRECTORY_PICTURES)
+        return File.createTempFile("PlantDiary${timestamp}", ".jpg", storageDir).apply {
+            currentPhotoPath = absolutePath
+        }
+    }
+
+    override fun onRequestPermissionsResult(
+        requestCode: Int,
+        permissions: Array<out String>,
+        grantResults: IntArray
+    ) {
+
+        when(requestCode) {
+            CAMERA_PERMISSION_REQUEST_CODE -> {
+                if (grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                    // permission granted, let's do stuff.
+                    takePhoto()
+                } else {
+                    Toast.makeText(context, "Unable to take photo without permission", Toast.LENGTH_LONG).show()
+                }
+            }
+            else -> {
+                super.onRequestPermissionsResult(requestCode, permissions, grantResults)
+            }
+
+        }
+    }
+}
