+++ app/src/main/java/edu/uc/jonesbr/myplantdiary/ui/main/MainFragment.kt
@@ -32,21 +32,16 @@ import java.io.File
 import java.text.SimpleDateFormat
 import java.util.*
 
-class MainFragment : Fragment() {
+class MainFragment : DiaryFragment() {
 
     private val IMAGE_GALLERY_REQUEST_CODE: Int = 2001
-    private val SAVE_IMAGE_REQUEST_CODE: Int = 1999
-    private val CAMERA_REQUEST_CODE: Int = 1998
-    private val CAMERA_PERMISSION_REQUEST_CODE = 1997
     private val LOCATION_PERMISSION_REQUEST_CODE = 2000
     private val AUTH_REQUEST_CODE = 2002
-    private lateinit var currentPhotoPath: String
     private lateinit var viewModel: MainViewModel
     private lateinit var locationViewModel: LocationViewModel
     private var _plantId = 0
     private var user : FirebaseUser? = null
     private var photos : ArrayList<Photo> = ArrayList<Photo>()
-    private var photoURI : Uri? = null
     private var specimen = Specimen()
 
     override fun onCreateView(
@@ -152,34 +147,12 @@ class MainFragment : Fragment() {
         }
     }
 
-    /**
-     * See if we have permission or not.
-     */
-    private fun prepTakePhoto() {
-        if (ContextCompat.checkSelfPermission(context!!, Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED) {
-            takePhoto()
-        } else {
-            val permissionRequest = arrayOf(Manifest.permission.CAMERA);
-            requestPermissions(permissionRequest, CAMERA_PERMISSION_REQUEST_CODE)
-
-        }
-    }
-
     override fun onRequestPermissionsResult(
         requestCode: Int,
         permissions: Array<out String>,
         grantResults: IntArray
     ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
         when(requestCode) {
-            CAMERA_PERMISSION_REQUEST_CODE -> {
-                if (grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    // permission granted, let's do stuff.
-                    takePhoto()
-                } else {
-                    Toast.makeText(context, "Unable to take photo without permission", Toast.LENGTH_LONG).show()
-                }
-            }
             LOCATION_PERMISSION_REQUEST_CODE -> {
                 if (grantResults.isNotEmpty() && grantResults[0] ==  PackageManager.PERMISSION_GRANTED) {
                     requestLocationUpdates()
@@ -187,22 +160,8 @@ class MainFragment : Fragment() {
                     Toast.makeText(context, "Unable to update location without permission", Toast.LENGTH_LONG).show()
                 }
             }
-        }
-    }
-
-    private fun takePhoto() {
-        Intent(MediaStore.ACTION_IMAGE_CAPTURE).also{
-            takePictureIntent -> takePictureIntent.resolveActivity(context!!.packageManager)
-            if (takePictureIntent == null) {
-                Toast.makeText(context, "Unable to save photo", Toast.LENGTH_LONG).show()
-            } else {
-                // if we are here, we have a valid intent.
-                val photoFile:File = createImageFile()
-                photoFile?.also {
-                    photoURI = FileProvider.getUriForFile(activity!!.applicationContext, "com.myplantdiary.android.fileprovider", it)
-                    takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, photoURI)
-                    startActivityForResult(takePictureIntent, SAVE_IMAGE_REQUEST_CODE)
-                }
+            else -> {
+                super.onRequestPermissionsResult(requestCode, permissions, grantResults)
             }
         }
     }
@@ -229,16 +188,6 @@ class MainFragment : Fragment() {
             } else if (requestCode == AUTH_REQUEST_CODE) {
                 user = FirebaseAuth.getInstance().currentUser
             }
-        }
-    }
-
-    private fun createImageFile() : File {
-        // genererate a unique filename with date.
-        val timestamp: String = SimpleDateFormat("yyyyMMdd_HHmmss").format(Date())
-        // get access to the directory where we can write pictures.
-        val storageDir:File? = context!!.getExternalFilesDir(Environment.DIRECTORY_PICTURES)
-        return File.createTempFile("PlantDiary${timestamp}", ".jpg", storageDir).apply {
-            currentPhotoPath = absolutePath
         }
     }
